+++ src/CMakeLists.txt
@@ -1,16 +1,45 @@
 # src/CMakeLists.txt
-# add sub-directories that need to declare targets
-add_subdirectory (sadmd)
-add_subdirectory (sadcd)
-# add_subdirectory (sadfsd)
-add_subdirectory (bootstrap)
+# common components that almost all targets depend on
+set (COMMON_SOURCE_FILES
+	comm/inet.cpp comm/socket.cpp)
 
-# common components that others depend on
-set (COMMON_SOURCES comm/inet.cpp comm/socket.cpp)
+# add main targets
+add_executable (sadmd
+	sadmd/sadmd.cpp
+	${COMMON_SOURCE_FILES})
+add_executable (sadcd
+	sadcd/sadcd.cpp
+	${COMMON_SOURCE_FILES})
+# add_executable (sadfsd
+#	sadfsd/sadfsd.cpp
+#	${COMMON_SOURCE_FILES})
 
-# declare dependencies for relevant targets
-target_sources (sadmd  PRIVATE ${COMMON_SOURCES})
-target_sources (sadcd  PRIVATE ${COMMON_SOURCES})
-# target_sources (sadfsd PRIVATE ${COMMON_SOURCES})
-target_sources (echod  PRIVATE ${COMMON_SOURCES})
-target_sources (msg    PRIVATE ${COMMON_SOURCES})
+# define a function to do the following for every main daemon target
+# 1. add a corresponding bootstrap target
+# 2. mark it as a dependency of the main target
+# 3. inform CMake of any external dependencies
+function (add_bootstrap TARGET)
+	set (BOOTSTRAP_TARGET ${TARGET}-bootstrap)
+	add_executable (${BOOTSTRAP_TARGET}
+		${TARGET}/${BOOTSTRAP_TARGET}.cpp bootstrap/util.cpp)
+	add_dependencies (${TARGET} ${BOOTSTRAP_TARGET})
+	target_include_directories (${BOOTSTRAP_TARGET}
+		PRIVATE ${Boost_INCLUDE_DIRS})
+	target_link_libraries (${BOOTSTRAP_TARGET}
+		PRIVATE ${Boost_LIBRARIES})
+endfunction()
+
+# first, find the boost_program_options library. this sets
+# the Boost_INCLUDE_DIRS and Boost_LIBRARIES variables
+find_package (Boost
+	1.67.0
+	REQUIRED
+	COMPONENTS program_options)
+
+add_bootstrap (sadmd)
+add_bootstrap (sadcd)
+# add_bootstrap (sadfsd)
+
+# add source files for "example" targets
+target_sources (echod  PRIVATE ${COMMON_SOURCE_FILES})
+target_sources (msg    PRIVATE ${COMMON_SOURCE_FILES})
