--- src/sadcd/CMakeLists.txt
@@ -1,2 +1,0 @@
-# src/sadcd/CMakeLists.txt
-add_executable (sadcd sadcd.cpp)
